Refactoring, cleaning up timeline markup
